requirements.txt
opencv-python
pillow
scikit-learn
PyYAML
easydict
matplotlib
tqdm
fontTools




